♻️ refactor: 구독 이력 서비스 인터페이스 및 구현체 추가
